Update CVCard.ts
if the string arrays such as language[0] or skill[0] have empty index positions in the 0th, the CV card will be hidden. However, SolidOS webapp doesn't seem to mind if the first index position where "English" may be, is allowed to be blank while another index position such as languages[1] where it is not blank, will display the CVCard. Maybe it would be good if a user had to fill out the first language they knew or they couldn't create multiple empty divs that were left blank. 

Left renderRoles() out from the TemplateResult, because at this time, the solidcommunity.net server is slow and unresponsive. Also, because the startDate and endDate interfere with how the string[] array is displayed. The array is technically not empty, but the user on SolidWeb App can make a startDate and endDate without filling out additional details. I'm positive if if it was left undefined on the backend or you could reset the startDate and endDate on the backend, you could hide the CVCard with renderRoles() easily.