Update CVPresenter.ts
Removed the weird skill question mark characters. Removed the underscore because it interfered with the renderLan() function in the CVCard's ability to be hidden. It is not needed and does not affect the functionality of how the languages are displayed.

Changed the slice to better reflect startDate and endDate. Any other changes that were made were suggested by VSCode.