Update QRCodeCard.ts
Shouldn't we change QRcode color to closer match the color scheme of Solid? Snapchat QR and Instagram QR have matching color schemes and distinctions for their codes as well. Changing the color does not seem to challenge or hurt a camera phones ability to absorb the information. I have already tested with my camera on the QRcode color scheme to make sure it can properly read the uri of the user. It would influence the user's interest in the aesthetic of UI/UX.