Update QRCodeCard.ts
This is a proof of concept that works for Android 12 and Android 14 phones. Your first name and URI should fill in the contacts list when scanned. the const name for URL can be changed to URI for naming conventions, just gotta make sure that the URL: in the string field is consistent with the rules of the vCard.