Update QRCodeCard.ts
Tested on Android 12 and Android 14 phones. First name and webid show up.

While subject.uri is a URI, that is up to you all if you wish to edit the const URL:string to URI portion for solid naming conventions/vocabulary. I left it alone out of the sake for vCard purposes. It may not make a difference to change things, but I left it up to whoever decides to make the revision.