Update QRCodeCard.ts
If you make an observation regarding the const URL: string, you'll notice I didn't change it to URI for either the string containing the subject.uri or the naming of the const itself. I wanted to follow vCard conventions. But if you feel that it should be changed more to fit with the solid ecosystem/vocan/naming conventions, go for it.